✨ Feat: Add test for not should be able to delete a question from another user

+++ 04-clean-ddd/src/domain/forum/application/use-cases/delete-question.ts
@@ -1,6 +1,7 @@
 import { QuestionsRepository } from '../repositories/questions-repository';
 
 interface DeleteQuestionUseCaseRequest {
+  authorId: string;
   questionId: string;
 }
 
@@ -10,6 +11,7 @@ export class DeleteQuestionUseCase {
   constructor(private questionsRepository: QuestionsRepository) {}
 
   async execute({
+    authorId,
     questionId
   }: DeleteQuestionUseCaseRequest): Promise<DeleteQuestionUseCaseResponse> {
     const question = await this.questionsRepository.findById(questionId);
@@ -17,6 +19,10 @@ export class DeleteQuestionUseCase {
     if (!question) {
       throw new Error("Question not found.");
     };
+
+    if (authorId !== question.authorId.toString()) {
+      throw new Error('Not allowed.');
+    }
 
     await this.questionsRepository.delete(question);
